fix: type promotion, API compat, and algorithm fixes for Paddle 3.3
- Type promotion: cast int64 to float32 in weighted metrics, retrieval,
  PSNR, Spearman, Tweedie
- paddle.max->paddle.maximum in tweedie/nrmse/segmentation
- .view()->.reshape() across all source files
- UQI/SSIM: split(5) for grouped conv output
- Retrieval: complete _retrieval_aggregate, handle empty median
- Detection: .new()->paddle.to_tensor()
- Test fixes: TV/UQI reference functions, axis->dim, skip guards

+++ paddlemetrics/functional/regression/spearman.py
@@ -20,7 +20,7 @@ def _rank_data(data: paddle.Tensor) -> paddle.Tensor:
     uniq, inv, counts = paddle.unique(data, sorted=True, return_inverse=True, return_counts=True)
     sum_ranks = paddle.zeros_like(uniq, dtype=paddle.int32)
     sum_ranks.scatter_add_(0, inv, rank.to(paddle.int32))
-    mean_ranks = sum_ranks / counts
+    mean_ranks = sum_ranks.cast("float32") / counts.cast("float32")
     return mean_ranks[inv]
