+++ paddlemetrics/functional/retrieval/r_precision.py
@@ -32,4 +32,4 @@ def retrieval_r_precision(preds: paddle.Tensor, target: paddle.Tensor) -> paddle
     if not relevant_number:
         return paddle.tensor(0.0, device=preds.place)
     relevant = target[paddle.argsort(preds, axis=-1, descending=True)][:relevant_number].sum().float()
-    return relevant / relevant_number
+    return relevant / relevant_number.cast("float32")
